+++ src/runtime/cmd/kata-runtime/kata-env_test.go
@@ -27,7 +27,7 @@ import (
 
 	"github.com/kata-containers/kata-containers/src/runtime/pkg/katatestutils"
 	"github.com/kata-containers/kata-containers/src/runtime/pkg/katautils"
-	"github.com/kata-containers/kata-containers/src/runtime/virtcontainers/pkg/oci"
+	"github.com/kata-containers/kata-containers/src/runtime/pkg/oci"
 	"github.com/stretchr/testify/assert"
 )
