+++ src/runtime/pkg/containerd-shim-v2/wait.go
@@ -17,7 +17,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
 
-	"github.com/kata-containers/kata-containers/src/runtime/virtcontainers/pkg/oci"
+	"github.com/kata-containers/kata-containers/src/runtime/pkg/oci"
 )
 
 const defaultCheckInterval = 1 * time.Second
