+++ src/runtime/virtcontainers/container.go
@@ -1,4 +1,6 @@
+//go:build linux
 // +build linux
+
 // Copyright (c) 2016 Intel Corporation
 // Copyright (c) 2014,2015,2016,2017 Docker, Inc.
 // SPDX-License-Identifier: Apache-2.0
