+++ src/runtime/virtcontainers/tap_endpoint.go
@@ -12,8 +12,8 @@ import (
 	"github.com/containernetworking/plugins/pkg/ns"
 	"github.com/vishvananda/netlink"
 
+	"github.com/kata-containers/kata-containers/src/runtime/pkg/uuid"
 	persistapi "github.com/kata-containers/kata-containers/src/runtime/virtcontainers/persist/api"
-	"github.com/kata-containers/kata-containers/src/runtime/virtcontainers/pkg/uuid"
 	vcTypes "github.com/kata-containers/kata-containers/src/runtime/virtcontainers/types"
 )
