+++ src/runtime/virtcontainers/utils/utils_linux_generic.go
@@ -1,3 +1,4 @@
+//go:build amd64 || arm64 || s390x || !ppc64le
 // +build amd64 arm64 s390x !ppc64le
 
 // Copyright (c) 2019 IBM
